fix minot review comments
