+++ packages/scalecube-discovery/.gitignore
@@ -10,3 +10,6 @@ yarn-error.log
 lib/
 es/
 
+
+report.cjs.html
+report.es.html
